change latitude and longitude from float to double

+++ src/OsmSharp/IO/PBF/Encoder.cs
@@ -659,7 +659,7 @@ namespace OsmSharp.IO.PBF
         /// Encodes a lat/lon value into an offset.
         /// </summary>
         /// <returns></returns>
-        public static long EncodeLatLon(float value, long offset, long granularity)
+        public static long EncodeLatLon(double value, long offset, long granularity)
         {
             return ((long)(value / .000000001) - offset) / granularity;
         }
